@@ -4,7 +4,7 @@ version: 1.6.2
 homepage: https://github.com/benjamindean/flutter_vibration
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
   flutter: ">=1.20.0 <2.0.0"
 
 dependencies:
